Fixed state machine bugs
Two bugs in the finite state machine that receives and parses packets were fixed in this release. These bugs could cause the state machine to corrupt memory if radio static occurred at precise moments during a transmission. The typical behavior would be that the robot would be working fine, then all the sudden just hang and become unresponsive until a reboot occurred. Also, in a radio noisy environment, it could make it difficult for the gamepad to initally control the robot.