Added support for digital servos
Digital servos work better at a higher pwm frequency. This code defaults to 120 hz but you can put a shunt between D5 and D6 on the nano to get it back down to 60 hz (safer for analog servos). Please note a lot of fake MG90s claim to be digital but are really analog.

This version also has code that detects when the servo controller has gone to sleep (usually due to a short power dip) and wakes it up automatically. This extends battery life and compensates for the occassional loose power wire.