widened range for demo mode
It turns out that if you turn the dial very close to the borderline of demo mode and RC mode, you can get into a situation where the robot bounced back and forth between modes and this can manifest as an apparent loss of power (especially if the gamepad is turned off). By widening the range, we make it much less likely that people who point the dial to DEMO will end up in this border situation. When turning to RC mode most people turn the dial all the way clockwise so that's not as much of an issue.