Cleanups and recording fixes
Several cleanups such as variable name consistency and reduction of debug output.

Revised version numbers to V2 and prefixed with gamepad "G" and robot "R" on debug output.

Also, had to change things so there is no attempt to record on gamepad at the same time as scratch longclick playback is in progress. There just isn't enough memory to keep two SD card files open at once. The logic is all still there, I just hacked the state machine to make sure both files won't be in use at the same time. It is possible to fix this by writing a special function to alternatively open and close files using a small buffer, but it's a lot of work and requires a ton of testing to make sure the SD card can respond fast enough, so deferring this work for this release.