Gamepad format sd card, robot minor changes
The main addition is the ability to format the SD card right inside the gamepad by booting the gamepad while holding down R4.

Pixy.h was commented out of the robot code for now, it wasn't being used anyway and for some reason a lot of people had trouble compiling with it there.