Pad packets for defective hc05
This version allows you to boot up while holding D2 to turn on pad-packet mode, and boot holding D1 to turn it off. The value is stored in EEPROM. This pads out radio packets as a work-around to some defective HC05 modules that have been flooding the market recently. They won't transmit until they receive a certain number of bytes, making the robot very laggy (2 to 3 second lag). The pad-out method reduces the lag to about 0.1 seconds which is acceptable.